Finalized dietitian code

+++ lib/features/dietitian_result_screen/data/repository/dietitian_result_repository.dart
@@ -23,9 +23,9 @@ import '../model/dietitian_result_model.dart';
 
 class DietitianResultRepository {
   Future<DietitianResultModel> fetchResults({
-    required int acetone,
-    required int ethanol,
-    required int hydrogen,
+    required double acetone,
+    required double ethanol,
+    required double hydrogen,
     required bool diabetic,
     required String goal,
     int debug = 1,
